requirements.txt
absl-py==0.7.1
astor==0.7.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
MarkupSafe==1.1.1
mock==3.0.5
nltk==3.2.4
numpy==1.16.3
pandas==0.24.2
Pillow==6.0.0
preprocessing==0.1.13
protobuf==3.7.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
scikit-learn==0.21.1
scipy==1.3.0
six==1.12.0
sphinx-rtd-theme==0.2.4
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.25.2
Werkzeug==0.15.4
